+++ modules/genealogy-collaboration-livewire/src/CollaborationLivewireServiceProvider.php
@@ -14,6 +14,10 @@ final class CollaborationLivewireServiceProvider extends ServiceProvider
         $this->loadViewsFrom(__DIR__.'/../resources/views', 'genealogy-collaboration-livewire');
         Livewire::component('genealogy-collaboration-list', CollaborationSpaceList::class);
         Livewire::component('module-genealogy-collaboration::collaboration-list', CollaborationSpaceList::class);
+        Livewire::component('module-genealogy-collaboration::proposal-list', CollaborationProposalList::class);
+        Livewire::component('module-genealogy-collaboration::proposal-editor', CollaborationProposalEditor::class);
+        Livewire::component('genealogy-collaboration-proposal-list', CollaborationProposalList::class);
+        Livewire::component('genealogy-collaboration-proposal-editor', CollaborationProposalEditor::class);
     }
 }